# backend/main.py
# from contextlib import asynccontextmanager
# import os
# import sys
# import asyncio
# from typing import AsyncGenerator, Optional

# from datetime import datetime, timedelta
# from uuid import UUID, uuid4

# from fastapi import FastAPI, HTTPException, Request, status, Depends
# from fastapi.responses import JSONResponse
# from fastapi.middleware.cors import CORSMiddleware
# from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
# from sqlmodel import SQLModel, Field, select
# from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
# from sqlalchemy.orm import sessionmaker
# from dotenv import load_dotenv
# from passlib.context import CryptContext
# from jose import JWTError, jwt

# # Fix for Windows Asyncio Loop with Psycopg
# if sys.platform == "win32":
#     asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())

# # Load environment variables
# load_dotenv()

# DATABASE_URL = os.getenv("DATABASE_URL")
# SECRET_KEY = os.getenv("SECRET_KEY")
# ALGORITHM = os.getenv("ALGORITHM", "HS256")
# ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30))

# pwd_context = CryptContext(schemes=["argon2"], deprecated="auto")
# oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/login")

# # T010: User Model
# class User(SQLModel, table=True):
#     __tablename__ = "user"
#     id: UUID = Field(default_factory=uuid4, primary_key=True)
#     email: str = Field(unique=True, index=True)
#     password_hash: str
#     name: Optional[str] = None
#     created_at: datetime = Field(default_factory=datetime.utcnow)

# class UserCreate(SQLModel):
#     email: str
#     password: str
#     name: Optional[str] = None

# class UserResponse(SQLModel):
#     id: UUID
#     email: str
#     name: Optional[str] = None
#     created_at: datetime

# class Token(SQLModel):
#     access_token: str
#     token_type: str

# # T015: Todo Model
# class Todo(SQLModel, table=True):
#     id: Optional[int] = Field(default=None, primary_key=True)
#     user_id: UUID = Field(foreign_key="user.id", index=True)
#     title: str = Field(max_length=100)
#     description: Optional[str] = Field(default=None, max_length=500)
#     completed: bool = Field(default=False)
#     created_at: datetime = Field(default_factory=datetime.utcnow)
#     updated_at: datetime = Field(default_factory=datetime.utcnow)

# # T016: Todo Schemas
# class TodoCreate(SQLModel):
#     title: str = Field(min_length=1, max_length=100)
#     description: Optional[str] = Field(default=None, max_length=500)

# class TodoUpdate(SQLModel):
#     title: Optional[str] = Field(default=None, min_length=1, max_length=100)
#     description: Optional[str] = Field(default=None, max_length=500)
#     completed: Optional[bool] = None

# class TodoResponse(SQLModel):
#     id: int
#     title: str
#     description: Optional[str]
#     completed: bool
#     created_at: datetime
#     updated_at: datetime

# # T011: Password Utils
# def verify_password(plain_password, hashed_password):
#     return pwd_context.verify(plain_password, hashed_password)

# def get_password_hash(password):
#     return pwd_context.hash(password)

# def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
#     to_encode = data.copy()
#     if expires_delta:
#         expire = datetime.utcnow() + expires_delta
#     else:
#         expire = datetime.utcnow() + timedelta(minutes=15)
#     to_encode.update({"exp": expire})
#     encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
#     return encoded_jwt
# # Create Async Engine
# # For development, use SQLite to avoid schema conflicts
# DATABASE_URL = "postgresql://neondb_owner:npg_mCBE5xGAs8iK@ep-sweet-salad-ah75xvt2-pooler.c-3.us-east-1.aws.neon.tech/neondb?sslmode=require&channel_binding=require"
# # if not DATABASE_URL:
# #     # Use a default fallback for local dev if not found, or raise warning
# #     # Note: This is crucial for local testing if .env is missing
# #     DATABASE_URL = "sqlite+aiosqlite:///database.db"
# # elif DATABASE_URL.startswith("postgresql://"):
# #     DATABASE_URL = DATABASE_URL.replace("postgresql://", "postgresql+psycopg://")

# engine = create_async_engine(DATABASE_URL, echo=True, future=True)

# async def get_session() -> AsyncGenerator[AsyncSession, None]:
#     async_session = sessionmaker(
#         engine, class_=AsyncSession, expire_on_commit=False
#     )
#     async with async_session() as session:
#         yield session

# async def get_current_user(token: str = Depends(oauth2_scheme), session: AsyncSession = Depends(get_session)):
#     credentials_exception = HTTPException(
#         status_code=status.HTTP_401_UNAUTHORIZED,
#         detail="Could not validate credentials",
#         headers={"WWW-Authenticate": "Bearer"},
#     )
#     try:
#         payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
#         user_id: str = payload.get("sub")
#         if user_id is None:
#             raise credentials_exception
#     except JWTError:
#         raise credentials_exception

#     try:
#         uuid_user_id = UUID(user_id)
#     except ValueError:
#         raise credentials_exception

#     result = await session.execute(select(User).where(User.id == uuid_user_id))
#     user = result.scalars().first()
#     if user is None:
#         raise credentials_exception
#     return user

# async def create_db_and_tables():
#     async with engine.begin() as conn:
#         # Create all tables with correct schema
#         await conn.run_sync(SQLModel.metadata.create_all)
# @asynccontextmanager
# async def lifespan(app: FastAPI):
#     # Startup
#     await create_db_and_tables()
#     yield
#     # Shutdown

# app = FastAPI(
#     title="Todo Backend API",
#     lifespan=lifespan
# )

# # CORS Configuration
# origins = [
#     "http://localhost:3000",
#     "http://localhost:3001",
# ]

# app.add_middleware(
#     CORSMiddleware,
#     allow_origins=origins,
#     allow_credentials=True,
#     allow_methods=["*"],
#     allow_headers=["*"],
# )

# # Centralized Error Handling
# @app.exception_handler(HTTPException)
# async def http_exception_handler(request: Request, exc: HTTPException):
#     return JSONResponse(
#         status_code=exc.status_code,
#         content={"detail": exc.detail},
#     )

# @app.exception_handler(Exception)
# async def global_exception_handler(request: Request, exc: Exception):
#     print(exc) # Log to console
#     return JSONResponse(
#         status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
#         content={"detail": "Internal Server Error"},
#     )

# @app.get("/health")
# async def health_check():
#     return {"status": "ok"}

# # Auth Endpoints
# @app.post("/auth/signup", response_model=UserResponse, status_code=status.HTTP_201_CREATED)
# async def signup(user_in: UserCreate, session: AsyncSession = Depends(get_session)):
#     # Check if email exists
#     result = await session.execute(select(User).where(User.email == user_in.email))
#     existing_user = result.scalars().first()
#     if existing_user:
#         raise HTTPException(
#             status_code=status.HTTP_400_BAD_REQUEST,
#             detail="Email already registered"
#         )

#     hashed_password = get_password_hash(user_in.password)
#     user = User(
#         email=user_in.email,
#         password_hash=hashed_password,
#         name=user_in.name
#     )
#     session.add(user)
#     await session.commit()
#     await session.refresh(user)
#     return user

# @app.post("/auth/login", response_model=Token)
# async def login(form_data: OAuth2PasswordRequestForm = Depends(), session: AsyncSession = Depends(get_session)):
#     # OAuth2PasswordRequestForm expects username, we treat email as username
#     result = await session.execute(select(User).where(User.email == form_data.username))
#     user = result.scalars().first()

#     if not user or not verify_password(form_data.password, user.password_hash):
#         raise HTTPException(
#             status_code=status.HTTP_401_UNAUTHORIZED,
#             detail="Incorrect email or password",
#             headers={"WWW-Authenticate": "Bearer"},
#         )

#     access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
#     access_token = create_access_token(
#         data={"sub": str(user.id)}, expires_delta=access_token_expires
#     )
#     return {"access_token": access_token, "token_type": "bearer"}

# # Todo Endpoints
# @app.post("/api/todos", response_model=TodoResponse, status_code=status.HTTP_201_CREATED)
# async def create_todo(
#     todo_in: TodoCreate,
#     current_user: User = Depends(get_current_user),
#     session: AsyncSession = Depends(get_session)
# ):
#     todo = Todo(
#         title=todo_in.title,
#         description=todo_in.description,
#         user_id=current_user.id
#     )
#     session.add(todo)
#     await session.commit()
#     await session.refresh(todo)
#     return todo

# @app.get("/api/todos", response_model=list[TodoResponse])
# async def list_todos(
#     current_user: User = Depends(get_current_user),
#     session: AsyncSession = Depends(get_session)
# ):
#     result = await session.execute(
#         select(Todo).where(Todo.user_id == current_user.id).order_by(Todo.created_at)
#     )
#     return result.scalars().all()

# @app.get("/api/todos/{id}", response_model=TodoResponse)
# async def get_todo(
#     id: int,
#     current_user: User = Depends(get_current_user),
#     session: AsyncSession = Depends(get_session)
# ):
#     result = await session.execute(
#         select(Todo).where(Todo.id == id, Todo.user_id == current_user.id)
#     )
#     todo = result.scalars().first()
#     if not todo:
#         raise HTTPException(status_code=404, detail="Todo not found")
#     return todo

# @app.put("/api/todos/{id}", response_model=TodoResponse)
# async def update_todo(
#     id: int,
#     todo_in: TodoUpdate,
#     current_user: User = Depends(get_current_user),
#     session: AsyncSession = Depends(get_session)
# ):
#     result = await session.execute(
#         select(Todo).where(Todo.id == id, Todo.user_id == current_user.id)
#     )
#     todo = result.scalars().first()
#     if not todo:
#         raise HTTPException(status_code=404, detail="Todo not found")

#     if todo_in.title is not None:
#         todo.title = todo_in.title
#     if todo_in.description is not None:
#         todo.description = todo_in.description
#     if todo_in.completed is not None:
#         todo.completed = todo_in.completed

#     todo.updated_at = datetime.utcnow()
#     session.add(todo)
#     await session.commit()
#     await session.refresh(todo)
#     return todo

# @app.delete("/api/todos/{id}", status_code=status.HTTP_204_NO_CONTENT)
# async def delete_todo(
#     id: int,
#     current_user: User = Depends(get_current_user),
#     session: AsyncSession = Depends(get_session)
# ):
#     result = await session.execute(
#         select(Todo).where(Todo.id == id, Todo.user_id == current_user.id)
#     )
#     todo = result.scalars().first()
#     if not todo:
#         raise HTTPException(status_code=404, detail="Todo not found")

#     await session.delete(todo)
#     await session.commit()
#     return None

# @app.patch("/api/todos/{id}/complete", response_model=TodoResponse)
# async def mark_complete(
#     id: int,
#     current_user: User = Depends(get_current_user),
#     session: AsyncSession = Depends(get_session)
# ):
#     result = await session.execute(
#         select(Todo).where(Todo.id == id, Todo.user_id == current_user.id)
#     )
#     todo = result.scalars().first()
#     if not todo:
#         raise HTTPException(status_code=404, detail="Todo not found")

#     todo.completed = not todo.completed
#     todo.updated_at = datetime.utcnow()
#     session.add(todo)
#     await session.commit()
#     await session.refresh(todo)
#     return todo




from contextlib import asynccontextmanager
import os
import sys
import asyncio
from typing import AsyncGenerator, Optional
from datetime import datetime, timedelta
from uuid import UUID, uuid4

from fastapi import FastAPI, HTTPException, Request, status, Depends
from fastapi.responses import JSONResponse
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm

from sqlmodel import SQLModel, Field, select
from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
from sqlalchemy.orm import sessionmaker

from dotenv import load_dotenv
from passlib.context import CryptContext
from jose import JWTError, jwt


# -------------------------
# Windows Async Fix
# -------------------------
if sys.platform == "win32":
    asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())

# -------------------------
# ENV
# -------------------------
load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")
SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30))

if not DATABASE_URL:
    raise RuntimeError("DATABASE_URL not set")

# Convert to async PostgreSQL
if DATABASE_URL.startswith("postgresql://"):
    DATABASE_URL = DATABASE_URL.replace(
        "postgresql://",
        "postgresql+psycopg://"
    )

# -------------------------
# Database
# -------------------------
engine = create_async_engine(
    DATABASE_URL,
    echo=True,
    future=True
)

async def get_session() -> AsyncGenerator[AsyncSession, None]:
    async_session = sessionmaker(
        engine,
        class_=AsyncSession,
        expire_on_commit=False
    )
    async with async_session() as session:
        yield session


# -------------------------
# Security
# -------------------------
pwd_context = CryptContext(schemes=["argon2"], deprecated="auto")
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/login")


def get_password_hash(password: str) -> str:
    return pwd_context.hash(password)


def verify_password(plain: str, hashed: str) -> bool:
    return pwd_context.verify(plain, hashed)


def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=15))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)


# -------------------------
# Models
# -------------------------
class User(SQLModel, table=True):
    __tablename__ = "users"  # FIXED (not reserved keyword)

    id: UUID = Field(default_factory=uuid4, primary_key=True)
    email: str = Field(unique=True, index=True)
    password_hash: str
    name: Optional[str] = None
    created_at: datetime = Field(default_factory=datetime.utcnow)


class UserCreate(SQLModel):
    email: str
    password: str
    name: Optional[str] = None


class UserResponse(SQLModel):
    id: UUID
    email: str
    name: Optional[str]
    created_at: datetime


class Token(SQLModel):
    access_token: str
    token_type: str


class Todo(SQLModel, table=True):
    id: Optional[int] = Field(default=None, primary_key=True)
    user_id: UUID = Field(foreign_key="users.id", index=True)
    title: str = Field(max_length=100)
    description: Optional[str] = Field(default=None, max_length=500)
    completed: bool = Field(default=False)
    created_at: datetime = Field(default_factory=datetime.utcnow)
    updated_at: datetime = Field(default_factory=datetime.utcnow)


class TodoCreate(SQLModel):
    title: str
    description: Optional[str] = None


class TodoUpdate(SQLModel):
    title: Optional[str] = None
    description: Optional[str] = None
    completed: Optional[bool] = None


class TodoResponse(SQLModel):
    id: int
    title: str
    description: Optional[str]
    completed: bool
    created_at: datetime
    updated_at: datetime


# -------------------------
# Auth Utils
# -------------------------
async def get_current_user(
    token: str = Depends(oauth2_scheme),
    session: AsyncSession = Depends(get_session)
):
    credentials_exception = HTTPException(
        status_code=401,
        detail="Invalid credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )

    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        user_id = payload.get("sub")
        if not user_id:
            raise credentials_exception
        user_uuid = UUID(user_id)
    except (JWTError, ValueError):
        raise credentials_exception

    result = await session.execute(
        select(User).where(User.id == user_uuid)
    )
    user = result.scalars().first()
    if not user:
        raise credentials_exception

    return user


# -------------------------
# DB Init
# -------------------------
async def create_db_and_tables():
    async with engine.begin() as conn:
        await conn.run_sync(SQLModel.metadata.create_all)


@asynccontextmanager
async def lifespan(app: FastAPI):
    await create_db_and_tables()
    yield


# -------------------------
# App
# -------------------------
app = FastAPI(
    title="Todo API",
    lifespan=lifespan
)

app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


# -------------------------
# Errors
# -------------------------
@app.exception_handler(HTTPException)
async def http_error(_, exc: HTTPException):
    return JSONResponse(status_code=exc.status_code, content={"detail": exc.detail})


@app.exception_handler(Exception)
async def global_error(_, exc: Exception):
    print(exc)
    return JSONResponse(status_code=500, content={"detail": "Internal Server Error"})


# -------------------------
# Routes
# -------------------------
@app.get("/health")
async def health():
    return {"status": "ok"}


# ---------- AUTH ----------
@app.post("/auth/signup", response_model=UserResponse, status_code=201)
async def signup(
    user_in: UserCreate,
    session: AsyncSession = Depends(get_session)
):
    result = await session.execute(
        select(User).where(User.email == user_in.email)
    )
    if result.scalars().first():
        raise HTTPException(400, "Email already registered")

    user = User(
        email=user_in.email,
        password_hash=get_password_hash(user_in.password),
        name=user_in.name
    )
    session.add(user)
    await session.commit()
    await session.refresh(user)
    return user


@app.post("/auth/login", response_model=Token)
async def login(
    form_data: OAuth2PasswordRequestForm = Depends(),
    session: AsyncSession = Depends(get_session)
):
    result = await session.execute(
        select(User).where(User.email == form_data.username)
    )
    user = result.scalars().first()

    if not user or not verify_password(form_data.password, user.password_hash):
        raise HTTPException(401, "Invalid email or password")

    token = create_access_token(
        {"sub": str(user.id)},
        timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    )
    return {"access_token": token, "token_type": "bearer"}


# ---------- TODOS ----------
@app.post("/api/todos", response_model=TodoResponse, status_code=201)
async def create_todo(
    todo_in: TodoCreate,
    user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_session)
):
    todo = Todo(
        title=todo_in.title,
        description=todo_in.description,
        user_id=user.id
    )
    session.add(todo)
    await session.commit()
    await session.refresh(todo)
    return todo


@app.get("/api/todos", response_model=list[TodoResponse])
async def list_todos(
    user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_session)
):
    result = await session.execute(
        select(Todo).where(Todo.user_id == user.id)
    )
    return result.scalars().all()


@app.put("/api/todos/{todo_id}", response_model=TodoResponse)
async def update_todo(
    todo_id: int,
    todo_in: TodoUpdate,
    user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_session)
):
    result = await session.execute(
        select(Todo).where(Todo.id == todo_id, Todo.user_id == user.id)
    )
    todo = result.scalars().first()
    if not todo:
        raise HTTPException(404, "Todo not found")

    for field, value in todo_in.model_dump(exclude_unset=True).items():
        setattr(todo, field, value)

    todo.updated_at = datetime.utcnow()
    await session.commit()
    await session.refresh(todo)
    return todo


@app.delete("/api/todos/{todo_id}", status_code=204)
async def delete_todo(
    todo_id: int,
    user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_session)
):
    result = await session.execute(
        select(Todo).where(Todo.id == todo_id, Todo.user_id == user.id)
    )
    todo = result.scalars().first()
    if not todo:
        raise HTTPException(404, "Todo not found")

    await session.delete(todo)
    await session.commit()