# specs/003-todo-backend-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend API
  version: 1.0.0
  description: REST API for Todo Application
paths:
  /auth/signup:
    post:
      summary: Register a new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: {type: string, format: email}
                password: {type: string, minLength: 8}
      responses:
        201:
          description: User created
        400:
          description: Email already registered

  /auth/login:
    post:
      summary: Login and get token
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                username: {type: string} # OAuth2 spec uses username for email
                password: {type: string}
      responses:
        200:
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: {type: string}
                  token_type: {type: string}

  /api/todos:
    get:
      summary: Get all todos for current user
      security: [{bearerAuth: []}]
      responses:
        200:
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoRead'
    post:
      summary: Create a new todo
      security: [{bearerAuth: []}]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      responses:
        201:
          description: Todo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoRead'

  /api/todos/{id}:
    get:
      summary: Get specific todo
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          schema: {type: integer}
          required: true
      responses:
        200:
          description: Todo details
        404:
          description: Not found (or not owned by user)
    put:
      summary: Update todo
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: integer}
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        200:
          description: Updated todo
    delete:
      summary: Delete todo
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: integer}
      responses:
        204:
          description: Deleted successfully

  /api/todos/{id}/complete:
    patch:
      summary: Toggle completion status
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: integer}
      responses:
        200:
          description: Updated status

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TodoCreate:
      type: object
      required: [title]
      properties:
        title: {type: string}
        description: {type: string}
    TodoUpdate:
      type: object
      properties:
        title: {type: string}
        description: {type: string}
        completed: {type: boolean}
    TodoRead:
      type: object
      properties:
        id: {type: integer}
        title: {type: string}
        description: {type: string}
        completed: {type: boolean}
        created_at: {type: string, format: date-time}